Adding a threading_helper in test.support
